+++ src/commands/general/dfd.ts
@@ -1,9 +1,9 @@
-import { CommandDefinition } from '../../lib/command';
+import { CommandDefinition, replyWithMsg } from '../../lib/command';
 import { CommandCategory } from '../../constants';
 
 export const dfd: CommandDefinition = {
     name: ['dfd'],
     description: 'Provides link to the Digital Flight Dynamics Discord server',
     category: CommandCategory.GENERAL,
-    executor: (msg) => msg.channel.send('https://discord.gg/dfd'),
+    executor: (msg) => replyWithMsg(msg, 'https://discord.gg/dfd'),
 };