+++ src/commands/general/headwind.ts
@@ -1,9 +1,9 @@
-import { CommandDefinition } from '../../lib/command';
+import { CommandDefinition, replyWithMsg } from '../../lib/command';
 import { CommandCategory } from '../../constants';
 
 export const headwind: CommandDefinition = {
     name: ['headwind', 'hw'],
     description: 'Provides link to the Headwind Discord server',
     category: CommandCategory.GENERAL,
-    executor: (msg) => msg.channel.send('https://discord.gg/SYanvKEX7B'),
+    executor: (msg) => replyWithMsg(msg, 'https://discord.gg/SYanvKEX7B'),
 };